feat(earn-filter): move earn filter and add mobile scene
