+++ packages/blockchain-wallet-v4-frontend/src/scenes/Earn/Earn.types.ts
@@ -1,7 +1,5 @@
-import { ChangeEvent } from 'react'
-
 import { EarnEDDStatus, RewardsRatesType, StakingRatesType } from '@core/types'
-import { EarnTabsType, UserDataType } from 'data/types'
+import { UserDataType } from 'data/types'
 
 export type SuccessStateType = {
   earnEDDStatus: EarnEDDStatus
@@ -13,13 +11,4 @@ export type SuccessStateType = {
 
 export type HTMLElementEvent<T extends HTMLElement> = Event & {
   target: T
-}
-
-export type EarnFilterPropsType = {
-  earnTab: EarnTabsType
-  handleAssetClick: (status: boolean) => void
-  handleHistoryClick: () => void
-  handleSearch: (e: ChangeEvent<HTMLInputElement>) => void
-  handleTabClick: (tab: string) => void
-  showAvailableAssets: boolean
 }
