+++ packages/blockchain-wallet-v4-frontend/src/scenes/Earn/Filter/Filter.types.ts
@@ -1,0 +1,12 @@
+import { ChangeEvent } from 'react'
+
+import { EarnTabsType } from 'data/types'
+
+export type EarnFilterPropsType = {
+  earnTab: EarnTabsType
+  handleAssetClick: (status: boolean) => void
+  handleHistoryClick: () => void
+  handleSearch: (e: ChangeEvent<HTMLInputElement>) => void
+  handleTabClick: (tab: string) => void
+  showAvailableAssets: boolean
+}
