+++ packages/blockchain-wallet-v4-frontend/src/scenes/Earn/index.tsx
@@ -12,8 +12,8 @@ import { debounce } from 'utils/helpers'
 import Loading from './Earn.loading.template'
 import { CustomSceneWrapper, EarnContainer, Overlay } from './Earn.model'
 import getData from './Earn.selectors'
-import EarnFilter from './Earn.template.filter'
 import EarnHeader from './Earn.template.header'
+import EarnFilter from './Filter'
 import Message from './Message'
 import Table from './Table'
